--- a/x/cdp/abci_test.go
+++ b/x/cdp/abci_test.go
@@ -1,6 +1,7 @@
 package cdp_test
 
 import (
+	"fmt"
 	"math/rand"
 	"testing"
 	"time"
@@ -21,12 +22,10 @@ import (
 type ModuleTestSuite struct {
 	suite.Suite
 
-	keeper       cdp.Keeper
-	addrs        []sdk.AccAddress
-	app          app.TestApp
-	cdps         cdp.CDPs
-	ctx          sdk.Context
-	liquidations liquidationTracker
+	keeper cdp.Keeper
+	addrs  []sdk.AccAddress
+	app    app.TestApp
+	ctx    sdk.Context
 }
 
 type liquidationTracker struct {
@@ -39,56 +38,31 @@ func (suite *ModuleTestSuite) SetupTest() {
 	tApp := app.NewTestApp()
 	ctx := tApp.NewContext(true, abci.Header{Height: 1, Time: tmtime.Now()})
 	coins := []sdk.Coins{}
-	tracker := liquidationTracker{}
 
-	for j := 0; j < 100; j++ {
+	const numAddrs = 100
+	for j := 0; j < numAddrs; j++ {
 		coins = append(coins, cs(c("btc", 100000000), c("xrp", 10000000000)))
 	}
-	_, addrs := app.GeneratePrivKeyAddressPairs(100)
+	_, addrs := app.GeneratePrivKeyAddressPairs(numAddrs)
 
-	authGS := app.NewAuthGenState(
-		addrs, coins)
 	tApp.InitializeFromGenesisStates(
-		authGS,
+		app.NewAuthGenState(addrs, coins),
 		NewPricefeedGenStateMulti(),
 		NewCDPGenStateMulti(),
 	)
 	suite.ctx = ctx
 	suite.app = tApp
 	suite.keeper = tApp.GetCDPKeeper()
-	suite.cdps = cdp.CDPs{}
 	suite.addrs = addrs
-	suite.liquidations = tracker
 }
 
-func (suite *ModuleTestSuite) createCdps() {
-	tApp := app.NewTestApp()
-	ctx := tApp.NewContext(true, abci.Header{Height: 1, Time: tmtime.Now()})
-	cdps := make(cdp.CDPs, 100)
-	_, addrs := app.GeneratePrivKeyAddressPairs(100)
-	coins := []sdk.Coins{}
+func createCDPs(ctx sdk.Context, keeper cdp.Keeper, addrs []sdk.AccAddress) (liquidationTracker, error) {
 	tracker := liquidationTracker{}
 
 	for j := 0; j < 100; j++ {
-		coins = append(coins, cs(c("btc", 100000000), c("xrp", 10000000000)))
-	}
+		var collateral string
+		var amount, debt int
 
-	authGS := app.NewAuthGenState(
-		addrs, coins)
-	tApp.InitializeFromGenesisStates(
-		authGS,
-		NewPricefeedGenStateMulti(),
-		NewCDPGenStateMulti(),
-	)
-
-	suite.ctx = ctx
-	suite.app = tApp
-	suite.keeper = tApp.GetCDPKeeper()
-
-	for j := 0; j < 100; j++ {
-		collateral := "xrp"
-		amount := 10000000000
-		debt := simulation.RandIntBetween(rand.New(rand.NewSource(int64(j))), 750000000, 1249000000)
 		if j%2 == 0 {
 			collateral = "btc"
 			amount = 100000000
@@ -98,67 +72,76 @@ func (suite *ModuleTestSuite) createCdps() {
 				tracker.debt += int64(debt)
 			}
 		} else {
+			collateral = "xrp"
+			amount = 10000000000
+			debt = simulation.RandIntBetween(rand.New(rand.NewSource(int64(j))), 750000000, 1249000000)
 			if debt >= 1000000000 {
 				tracker.xrp = append(tracker.xrp, uint64(j+1))
 				tracker.debt += int64(debt)
 			}
 		}
-		suite.Nil(suite.keeper.AddCdp(suite.ctx, addrs[j], c(collateral, int64(amount)), c("usdx", int64(debt)), collateral+"-a"))
-		c, f := suite.keeper.GetCDP(suite.ctx, collateral+"-a", uint64(j+1))
-		suite.True(f)
-		cdps[j] = c
+		err := keeper.AddCdp(ctx, addrs[j], c(collateral, int64(amount)), c("usdx", int64(debt)), collateral+"-a")
+		if err != nil {
+			return liquidationTracker{}, err
+		}
+		_, f := keeper.GetCDP(ctx, collateral+"-a", uint64(j+1))
+		if !f {
+			return liquidationTracker{}, fmt.Errorf("created cdp, but could not find in store")
+		}
 	}
-
-	suite.cdps = cdps
-	suite.addrs = addrs
-	suite.liquidations = tracker
+	return tracker, nil
 }
 
 func (suite *ModuleTestSuite) setPrice(price sdk.Dec, market string) {
 	pfKeeper := suite.app.GetPriceFeedKeeper()
 
-	pfKeeper.SetPrice(suite.ctx, sdk.AccAddress{}, market, price, suite.ctx.BlockTime().Add(time.Hour*3))
-	err := pfKeeper.SetCurrentPrices(suite.ctx, market)
+	_, err := pfKeeper.SetPrice(suite.ctx, sdk.AccAddress{}, market, price, suite.ctx.BlockTime().Add(time.Hour*3))
+	suite.NoError(err)
+	err = pfKeeper.SetCurrentPrices(suite.ctx, market)
 	suite.NoError(err)
 	pp, err := pfKeeper.GetCurrentPrice(suite.ctx, market)
 	suite.NoError(err)
 	suite.Equal(price, pp.Price)
 }
 func (suite *ModuleTestSuite) TestBeginBlock() {
-	suite.createCdps()
+	liquidations, err := createCDPs(suite.ctx, suite.keeper, suite.addrs)
+	suite.Require().NoError(err)
+
 	sk := suite.app.GetSupplyKeeper()
 	acc := sk.GetModuleAccount(suite.ctx, cdp.ModuleName)
 	originalXrpCollateral := acc.GetCoins().AmountOf("xrp")
+	originalBtcCollateral := acc.GetCoins().AmountOf("btc")
+
 	suite.setPrice(d("0.2"), "xrp:usd")
+	suite.setPrice(d("6000"), "btc:usd")
 	cdp.BeginBlocker(suite.ctx, abci.RequestBeginBlock{Header: suite.ctx.BlockHeader()}, suite.keeper)
+
 	acc = sk.GetModuleAccount(suite.ctx, cdp.ModuleName)
 	finalXrpCollateral := acc.GetCoins().AmountOf("xrp")
 	seizedXrpCollateral := originalXrpCollateral.Sub(finalXrpCollateral)
 	xrpLiquidations := int(seizedXrpCollateral.Quo(i(10000000000)).Int64())
-	suite.Equal(len(suite.liquidations.xrp), xrpLiquidations)
+	suite.Equal(len(liquidations.xrp), xrpLiquidations)
 
-	acc = sk.GetModuleAccount(suite.ctx, cdp.ModuleName)
-	originalBtcCollateral := acc.GetCoins().AmountOf("btc")
-	suite.setPrice(d("6000"), "btc:usd")
-	cdp.BeginBlocker(suite.ctx, abci.RequestBeginBlock{Header: suite.ctx.BlockHeader()}, suite.keeper)
 	acc = sk.GetModuleAccount(suite.ctx, cdp.ModuleName)
 	finalBtcCollateral := acc.GetCoins().AmountOf("btc")
 	seizedBtcCollateral := originalBtcCollateral.Sub(finalBtcCollateral)
 	btcLiquidations := int(seizedBtcCollateral.Quo(i(100000000)).Int64())
-	suite.Equal(len(suite.liquidations.btc), btcLiquidations)
+	suite.Equal(len(liquidations.btc), btcLiquidations)
 
 	acc = sk.GetModuleAccount(suite.ctx, auction.ModuleName)
-	suite.Equal(suite.liquidations.debt, acc.GetCoins().AmountOf("debt").Int64())
-
+	suite.Equal(liquidations.debt, acc.GetCoins().AmountOf("debt").Int64())
 }
 
 func (suite *ModuleTestSuite) TestSeizeSingleCdpWithFees() {
 	err := suite.keeper.AddCdp(suite.ctx, suite.addrs[0], c("xrp", 10000000000), c("usdx", 1000000000), "xrp-a")
 	suite.NoError(err)
+
 	suite.Equal(i(1000000000), suite.keeper.GetTotalPrincipal(suite.ctx, "xrp-a", "usdx"))
+
 	sk := suite.app.GetSupplyKeeper()
 	cdpMacc := sk.GetModuleAccount(suite.ctx, cdp.ModuleName)
 	suite.Equal(i(1000000000), cdpMacc.GetCoins().AmountOf("debt"))
+
 	for i := 0; i < 100; i++ {
 		suite.ctx = suite.ctx.WithBlockTime(suite.ctx.BlockTime().Add(time.Second * 6))
 		cdp.BeginBlocker(suite.ctx, abci.RequestBeginBlock{Header: suite.ctx.BlockHeader()}, suite.keeper)